Refactor artifacts CLI to new cli.py module
