+++ workflows/prognostic_run_diags/fv3net/artifacts/query.py
@@ -1,5 +1,4 @@
 from typing import Optional, Container, List, Awaitable, Sequence
-import argparse
 import asyncio
 import dataclasses
 import datetime
@@ -9,7 +8,6 @@ import pathlib
 import fsspec
 import gcsfs
 
-from .report_search import ReportIndex
 from .utils import _list, _close_session
 
 
@@ -120,17 +118,7 @@ def list(args):
             )
 
 
-def report_entrypoint(args):
-    if args.write:
-        index = ReportIndex()
-        index.compute(args.reports_url)
-        index.dump(os.path.join(args.reports_url, "index.json"))
-    index = ReportIndex.from_json(os.path.join(args.reports_url, "index.json"))
-    for link in index.public_links(args.url):
-        print(link)
-
-
-def register_ls_parser(subparsers):
+def register_parser(subparsers):
     parser = subparsers.add_parser("ls", help="Query the experiments buckets.")
     parser.add_argument("step", nargs="*", help="One of " + ", ".join(STEPS))
     parser.add_argument(
@@ -154,40 +142,3 @@ def register_ls_parser(subparsers):
         default=None,
     )
     parser.set_defaults(func=list)
-
-
-def register_report_parser(subparsers):
-    parser = subparsers.add_parser("report", help="Search for prognostic run reports.")
-    parser.add_argument("url", help="A prognostic run URL.")
-    parser.add_argument(
-        "-r",
-        "--reports-url",
-        help=(
-            "Location of prognostic run reports. Defaults to gs://vcm-ml-public/argo. "
-            "Search uses index at REPORTS_URL/index.json"
-        ),
-        default="gs://vcm-ml-public/argo",
-    )
-    parser.add_argument(
-        "-w",
-        "--write",
-        help="Recompute index and write to REPORTS_URL/index.json before searching.",
-        action="store_true",
-    )
-    parser.set_defaults(func=report_entrypoint)
-
-
-def get_parser():
-    parser = argparse.ArgumentParser(
-        description="Query available experiment and report output."
-    )
-    subparsers = parser.add_subparsers(required=True, dest="command")
-    register_ls_parser(subparsers)
-    register_report_parser(subparsers)
-    return parser
-
-
-def main():
-    parser = get_parser()
-    args = parser.parse_args()
-    args.func(args)
