+++ workflows/prognostic_run_diags/setup.py
@@ -47,7 +47,7 @@ setup(
     entry_points={
         "console_scripts": [
             "prognostic_run_diags=fv3net.diagnostics.prognostic_run.cli:main",
-            "artifacts=fv3net.artifacts.query:main",
+            "artifacts=fv3net.artifacts.cli:main",
         ]
     },
     install_requires=install_requirements,
